Lay groundwork for modularizing files containing routes
